requirements.txt
certifi==2018.1.18
chardet==3.0.4
fuzzywuzzy==0.16.0
idna==2.6
iniherit==0.3.9
pyTelegramBotAPI==3.6.2
python-Levenshtein==0.12.0
python-slugify==1.2.5
requests==2.18.4
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
yoyo-migrations==5.0.5
